Fix python log time

+++ openc3/python/openc3/utilities/logger.py
@@ -196,7 +196,7 @@ class Logger(metaclass=LoggerMeta):
         with self.instance_mutex:
             now_time = datetime.now(timezone.utc)
             data = {
-                "time": now_time.timestamp() * 1000000000,
+                "time": int(now_time.timestamp() * 1000000000),
                 # Can't use isoformat because it appends "+00:00" instead of "Z"
                 "@timestamp": now_time.strftime("%Y-%m-%dT%H:%M:%S.%fZ"),
                 "level": log_level,
